This is the 5.4 version of the fix for bug#47007    Unresolved reference to 'innodb_system_libs' in "mysql_config"
In 5.4.2, we use InnoDB 1.0.4 which does file IO via separate
threads, opposed to the use of asynchronous IO previously.

So there is no InnoDB call to "aio_read()" which was searched
in "librt", causing a "-lrt" value of "innodb_system_libs",
that whole variable is gone.

This fix was applied in the build of 5.4.2-beta.

scripts/Makefile.am:
  There is no "innodb_system_libs" variable any more,
  so it cannot be replaced by its value.
scripts/mysql_config.pl.in:
  InnoDB does not need any platform-specific libraries any more,
  "innodb_system_libs" may go.
scripts/mysql_config.sh:
  InnoDB does not need any platform-specific libraries any more,
  "innodb_system_libs" may go.