bug#47037 to port Bug@25998 fixes to telco 6.2
the 2nd patch to address 6.2 -> 6.3 possible upgrade issue.

The added ignore_server_ids param is stored as the last line in master.info on 6.2, but on 6.3 it would be buried by some more lines that are last currently  in there.

It was suggested to make the 6.2 configuration line to have the same offset
(the 18th) in 6.2 as it would be in 6.3.
Fixed with implementing the suggestion. The 6.2 master.info will have two blank lines before the terminal one describing ignore_server_ids.